+++ src/Puml/Generator.fs
@@ -160,7 +160,7 @@ module Generate =
             | PostEvent { Caller = caller; Stream = stream } ->
                 let event =
                     match stream with
-                    | Stream { StreamType = (DomainType (ResolvedType.Stream { EventType = TypeName event })) } -> event
+                    | Stream { StreamType = DomainType.Stream event } -> event
                     | participant -> failwithf "[Logic] There is no stream in the post event, but there is a %A" participant
 
                 let postEvent =
@@ -172,6 +172,23 @@ module Generate =
 
                 [
                     postEvent |> PumlPart.indent indentation
+                ]
+
+            | ReadEvent { Caller = caller; Stream = stream } ->
+                let event =
+                    match stream with
+                    | Stream { StreamType = DomainType.Stream event } -> event
+                    | participant -> failwithf "[Logic] There is no stream in the read event, but there is a %A" participant
+
+                let readEvent =
+                    sprintf "%s ->> %s: %s"
+                        (stream |> ActiveParticipant.name)
+                        (caller |> ActiveParticipant.name)
+                        event
+                    |> PumlPart
+
+                [
+                    readEvent |> PumlPart.indent indentation
                 ]
 
             | HandleEventInStream { Stream = stream; Service = service; Method = method; Execution = execution } ->