+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Parse left and right notes in tuc
+- Add tuc part
+    - `ReadEvent` from stream
 
 ## 0.1.0 - 2020-08-12
 - Initial implementation